expose static methods on main type
This is done for usability reasons, and allow users to write
Person.findMany instead of Chisel.Person.findMany.

The drawback is that users have to extend the class, but that
is a common pattern and we believe that to be okay.

Also note that the tests don't have to be changed. That's because
from an implementation PoV, we still rely on Chisel.Person being
present. Documentation, on the other hand, is changed. That includes
the types. If users don't extend the types, they can still access
Chisel.Type.

(I am happy to change tests, but only after we're agreed on this)