get rid of Chisel.save
We now have all the machinery to make this happen, so do it.

To save an object, construct it with its constructor in pure
typescript (or get an existing from the database), and then call .save()

Note that the test we had for writing arbitrary types to Chisel.save is
gone, because that's now simply not possible. Actually, tt would be
possible to try to write to the OAuthType, but *right now* it is not
possible because that object is not defined in the endpoint code, so the
save() method won't even be available. When we move method resolution
fully to the client it will be, at which point we can reintroduce it.

Follow up: notice that in the store.ts example, we now unfortunately
have to do a lot more work to get the object from Json. Creating a
typescript object from Json is really hard because the types are not
visible at runtime. However, we could use an introspection API and
provide a really cool way to do that.