Switch to esbuild instead of webpack
I propose we do this before we invest a lot more heavily into webpack.
I played with esbuild today and it looks a lot smoother. It doesn't
need a configuration file, was 10x easier to get started, plus it is 10x
faster. One interesting thing about it is that it *does not* call tsc.
That means that for users on vscode, we can disable that (disabled by
default) and now compiling 3 relatively big endpoints was a sub-second
process (the endpoints were big because all of them imported lodash
independently). The overhead at this point is the overhead of launching
the process.

The resulting bundled code is smaller, because we exclude and
assume-present the chiselstrike API (which is possible with webpack too,
it was just a lot easier to do with esbuild because the help literally
fits in the screen).