implement uniqueness constraints
As much as I don't like the idea of decorators, I found no other
solution that was satisfactory given TypeScript's limited type system:

There is a builtin syntax that looked promising called "Template literal
types" that allows you to write:

    type QuietGreeting = Lowercase<Greeting>

But:
1) I didn't find any indication that this is widely used, on the
   contrary
2) If we are to actually implement it, it would be by doing something
   like type Foo = Unique<T>. However I didn't find any indication that
   we would be able to do something useful with it. Functions that I
   tested that were supposed to take a Unique<T> were happy to take a T
   as well on TypeScript playground.

However, reviewers: The time for us to find a good syntax is *NOW* so
suggestions welcome.

Changing the syntax would be relatively simple, as most of the PR is
internal machinery.

The biggest change is that we need to add a uniqueness field into the
meta database, which means older versions of ChiselStrike would have
issues upgrading. To fix that I am adding a evolution system that should
allow us to keep evolving the meta database safely in the future. (with
tests)